Move tests with configuration to dedicated package
