+++ app/src/test/java/com/mysugr/android/testing/example/moduleconfig/FeatureRunner.kt
@@ -1,4 +1,4 @@
-package com.mysugr.android.testing.example
+package com.mysugr.android.testing.example.moduleconfig
 
 import cucumber.api.CucumberOptions
 import cucumber.api.junit.Cucumber
@@ -9,5 +9,5 @@ import org.junit.runner.RunWith
 @CucumberOptions(
         plugin = ["pretty"],
         features = ["src/test/resources/features"],
-        glue = ["com.mysugr.android.testing.example.feature"])
+        glue = ["com.mysugr.android.testing.example.moduleconfig"])
 class FeatureRunner
