+++ app/src/test/java/com/mysugr/android/testing/example/moduleconfig/auth/LoginIntegrationTestSteps.kt
@@ -1,11 +1,11 @@
-package com.mysugr.android.testing.example.feature.auth
+package com.mysugr.android.testing.example.moduleconfig.auth
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
+import com.mysugr.android.testing.example.moduleconfig.appModuleTestingConfiguration
 import com.mysugr.android.testing.example.auth.AuthManager
-import com.mysugr.android.testing.example.net.BackendGatewaySteps
-import com.mysugr.android.testing.example.state.SessionStoreSteps
+import com.mysugr.android.testing.example.moduleconfig.net.BackendGatewaySteps
+import com.mysugr.android.testing.example.moduleconfig.state.SessionStoreSteps
 import com.mysugr.android.testing.example.view.LoginViewModel
-import com.mysugr.android.testing.example.view.LoginViewModelSteps
+import com.mysugr.android.testing.example.moduleconfig.view.LoginViewModelSteps
 import com.mysugr.sweetest.framework.base.BaseSteps
 import com.mysugr.sweetest.framework.base.steps
 import com.mysugr.sweetest.framework.context.TestContext
