+++ app/src/test/java/com/mysugr/android/testing/example/moduleconfig/coroutine/TriggerAwaiterSteps.kt
@@ -1,6 +1,7 @@
-package com.mysugr.android.testing.example.coroutine
+package com.mysugr.android.testing.example.moduleconfig.coroutine
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
+import com.mysugr.android.testing.example.coroutine.TriggerAwaiter
+import com.mysugr.android.testing.example.moduleconfig.appModuleTestingConfiguration
 import com.mysugr.sweetest.framework.base.BaseSteps
 import com.mysugr.sweetest.framework.context.TestContext
 import com.mysugr.sweetest.framework.coroutine.throwExceptionIfFailed
