+++ app/src/test/java/com/mysugr/android/testing/example/moduleconfig/state/SessionStoreSteps.kt
@@ -1,8 +1,9 @@
-package com.mysugr.android.testing.example.state
+package com.mysugr.android.testing.example.moduleconfig.state
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
+import com.mysugr.android.testing.example.moduleconfig.appModuleTestingConfiguration
 import com.mysugr.android.testing.example.user.User
-import com.mysugr.android.testing.example.feature.auth.UserSteps
+import com.mysugr.android.testing.example.moduleconfig.auth.UserSteps
+import com.mysugr.android.testing.example.state.SessionStore
 import com.mysugr.sweetest.framework.base.*
 import com.mysugr.sweetest.framework.context.TestContext
