+++ app/src/test/java/com/mysugr/android/testing/example/moduleconfig/view/LoginViewModelTest.kt
@@ -1,8 +1,9 @@
-package com.mysugr.android.testing.example.view
+package com.mysugr.android.testing.example.moduleconfig.view
 
-import com.mysugr.android.testing.example.appModuleTestingConfiguration
-import com.mysugr.android.testing.example.auth.AuthManagerSteps
-import com.mysugr.android.testing.example.feature.auth.UserSteps
+import com.mysugr.android.testing.example.moduleconfig.appModuleTestingConfiguration
+import com.mysugr.android.testing.example.moduleconfig.auth.AuthManagerSteps
+import com.mysugr.android.testing.example.moduleconfig.auth.UserSteps
+import com.mysugr.android.testing.example.view.LoginViewModel
 import com.mysugr.sweetest.framework.base.*
 import org.junit.Test
